Move effector_popup into client and use worldspace for rendering
