+++ assets/typst/perf_metrics.typ
@@ -4,11 +4,13 @@
   #set text(fill: base7)
 
   #place(bottom + right)[
-    #box(fill: base1.transparentize(20%), outset: 10pt)[
-      #align(left)[
-        = Performance Metrics
+    #box(inset: 20pt)[
+      #rect(fill: base1.transparentize(20%), outset: 10pt)[
+        #align(left)[
+          = Performance Metrics
 
-        *FPS*: #fps
+          *FPS*: #fps
+        ]
       ]
     ]
   ]
